chore: Update release process.
As long as the private SSH key is within our CircleCI project settings, we do not need to specify the ssh key fingerprint.  I am unsure what the other fingerprint was for in the "release" job, but if that fingerprint is needed, consider adding the private key to the project settings.